// src/api/models/APIClassmate.ts
export type APIClassmate = {
  email: string;
  first_name: string;
  hours: number;
  is_online: boolean;
  is_tutor: boolean;
  joined: string;
  last_name: string;
  profile_image_url: string;
  rank: number;
  registered: boolean;
  role: any;
  role_id: number;
  scholarship_points: number;
  school: string;
  school_id: number;
  state: string;
  user_id: number;
};